♻️ refactor: HomeUserCase 오타 수정 (userCase → useCase)
디렉토리, 파일명, 패키지, 클래스명 일괄 수정
